+++ src/hooks/index.ts
@@ -1,4 +1,4 @@
+export type { DeviceEvent, DeviceState } from './useDevice'
+export { useDevice } from './useDevice'
 export type { EventLogState } from './useEventLog'
 export { useEventLog } from './useEventLog'
-export type { GamepadEvent, GamepadState } from './useGamepad'
-export { useGamepad } from './useGamepad'